+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/register.json/post.js
@@ -1,5 +1,5 @@
 ﻿module.exports = function(client, callback) {
-	impress.security.register(client, function(err, user) {
+	security.register(client, function(err, user) {
 		if (user) client.context.data = { Result: "Ok" };
 		else client.context.data = { Result: "Error" };
 		callback();