+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/signIn.json/post.js
@@ -1,5 +1,5 @@
 ﻿module.exports = function(client, callback) {
-	impress.security.signIn(client, function(isSuccess) {
+	security.signIn(client, function(isSuccess) {
 		if (isSuccess) client.context.data = { Result: "Ok" };
 		else client.context.data = { Result: "Error" };
 		if (client.fields.loginForm) client.redirect("/");