+++ examples/copyContentToProjectFolder/applications/example/app/api/auth/signOut.json/post.js
@@ -1,4 +1,4 @@
 ﻿module.exports = function(client, callback) {
-	impress.security.signOut(client, callback);
+	security.signOut(client, callback);
 	callback({ Result: "Ok" })
 }