+++ examples/copyContentToProjectFolder/applications/example/app/examples/tools/serverHealth.json/get.js
@@ -1,8 +1,8 @@
 module.exports = function(client, callback) {
 
-	client.application.leak = client.application.leak || [];
-	var n = client.application.leak.length ? client.application.leak.length : 0;
-	client.application.leak.push(new Array(1024*(1024/n+1)).join("Hello node.js, Hello Impress Application Server "));
+	application.leak = application.leak || [];
+	var n = application.leak.length ? application.leak.length : 0;
+	application.leak.push(new Array(1024*(1024/n+1)).join("Hello node.js, Hello Impress Application Server "));
 
 	callback({
 		memory: process.memoryUsage()