+++ examples/copyContentToProjectFolder/applications/example/app/setup/apply.json/post.js
@@ -3,24 +3,20 @@
     var npmList = [ 'mongodb', 'memcached', 'mysql', 'mysql-utilities', 'nodemailer', 'geoip-lite', 'websocket' ],
     	npmChecked = client.fields.npmChecked.split(',');
 
-	impress.npm.load(impress.npm.config, function (err) {
-
-		impress.npm.on("log", function (message) {
+	npm.load(npm.config, function (err) {
+		npm.on("log", function (message) {
 			console.log(message);
 		});
-
 		for (var i = 0; i < npmList.length; i++) {
 			var npmName = npmList[i],
-				lib = impress.require(npmName);
+				lib = require(npmName);
 			if (!lib && npmChecked.indexOf(npmName) != -1) {
-				impress.npm.commands.install([npmName], function (err, data) {
+				npm.commands.install([npmName], function (err, data) {
 					if (err) console.log('npm error'.red);
 				});
 			}
 		}
-
 	});
-
 	callback();
 
 }